Filter NSStringLocalizedFormatKey from macstringsdict
Smartling breaks when attempting to process NSStringLocalizedFormatKey.
So we will filter those out for and for the time being assign automatically
asign them 'do not translate'.
