+++ webapp/src/main/java/com/box/l10n/mojito/service/assetExtraction/extractor/AssetPathToFilterConfigMapper.java
@@ -1,12 +1,7 @@
 package com.box.l10n.mojito.service.assetExtraction.extractor;
 
-import com.box.l10n.mojito.okapi.filters.AndroidFilter;
-import com.box.l10n.mojito.okapi.filters.CSVFilter;
-import com.box.l10n.mojito.okapi.filters.JSFilter;
-import com.box.l10n.mojito.okapi.filters.MacStringsFilter;
-import com.box.l10n.mojito.okapi.filters.MacStringsdictFilter;
-import com.box.l10n.mojito.okapi.filters.POFilter;
-import com.box.l10n.mojito.okapi.filters.XMLFilter;
+import com.box.l10n.mojito.okapi.filters.*;
+
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Set;
@@ -38,6 +33,7 @@ public class AssetPathToFilterConfigMapper {
      
     public static final String MACSTRINGS_FILTER_CONFIG_ID = MacStringsFilter.FILTER_CONFIG_ID + "-macStrings";
     public static final String MACSTRINGSDICT_FILTER_CONFIG_ID = MacStringsdictFilter.FILTER_CONFIG_ID + "-stringsdict";
+    public static final String MACSTRINGSDICT_FILTER_KEY_CONFIG_ID = MacStringsdictFilterKey.FILTER_CONFIG_ID + "-stringsdict";
     public static final String RESX_FILTER_CONFIG_ID = XMLFilter.FILTER_CONFIG_ID + "-resx";
     public static final String XTB_FILTER_CONFIG_ID = XMLFilter.FILTER_CONFIG_ID + "-xtb";
     public static final String JS_FILTER_CONFIG_ID = JSFilter.FILTER_CONFIG_ID + "-js";
